chore: rename `keyscmd` to `keyscli`
