+++ gno.land/cmd/gnokey/main.go
@@ -5,18 +5,18 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/gnolang/gno/gno.land/pkg/keyscmd"
+	"github.com/gnolang/gno/gno.land/pkg/keyscli"
 	"github.com/gnolang/gno/gnovm/pkg/gnoenv"
 	"github.com/gnolang/gno/tm2/pkg/commands"
 )
 
 func main() {
-	baseCfg := keyscmd.BaseOptions{
+	baseCfg := keyscli.BaseOptions{
 		Home:   gnoenv.HomeDir(),
 		Remote: "127.0.0.1:26657",
 	}
 
-	cmd := keyscmd.NewRootCmdWithBaseConfig(commands.NewDefaultIO(), baseCfg)
+	cmd := keyscli.NewRootCmdWithBaseConfig(commands.NewDefaultIO(), baseCfg)
 	if err := cmd.ParseAndRun(context.Background(), os.Args[1:]); err != nil {
 		_, _ = fmt.Fprintf(os.Stderr, "%+v\n", err)
