+++ gno.land/pkg/integration/testing_integration.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/gnolang/gno/gno.land/pkg/keyscmd"
+	"github.com/gnolang/gno/gno.land/pkg/keyscli"
 	"github.com/gnolang/gno/gnovm/pkg/gnoenv"
 	"github.com/gnolang/gno/tm2/pkg/bft/node"
 	"github.com/gnolang/gno/tm2/pkg/commands"
@@ -181,7 +181,7 @@ func setupGnolandTestScript(t *testing.T, txtarDir string) testscript.Params {
 				io := commands.NewTestIO()
 				io.SetOut(commands.WriteNopCloser(ts.Stdout()))
 				io.SetErr(commands.WriteNopCloser(ts.Stderr()))
-				cmd := keyscmd.NewRootCmd(io)
+				cmd := keyscli.NewRootCmd(io)
 
 				io.SetIn(strings.NewReader("\n")) // Inject empty password to stdin.
 				defaultArgs := []string{
