+++ gno.land/pkg/keyscli/common.go
@@ -1,4 +1,4 @@
-package keyscmd
+package keyscli
 
 type BaseOptions struct {
 	Home                  string
