+++ gno.land/pkg/keyscli/delete_test.go
@@ -1,4 +1,4 @@
-package keyscmd
+package keyscli
 
 import (
 	"strings"
