+++ gno.land/pkg/keyscli/export_test.go
@@ -1,4 +1,4 @@
-package keyscmd
+package keyscli
 
 import (
 	"fmt"
