+++ gno.land/pkg/keyscli/generate.go
@@ -1,4 +1,4 @@
-package keyscmd
+package keyscli
 
 import (
 	"context"
