+++ gno.land/pkg/keyscli/import_test.go
@@ -1,4 +1,4 @@
-package keyscmd
+package keyscli
 
 import (
 	"fmt"
