+++ gno.land/pkg/keyscli/list_test.go
@@ -1,4 +1,4 @@
-package keyscmd
+package keyscli
 
 import (
 	"testing"
