+++ gno.land/pkg/keyscli/maketx.go
@@ -1,4 +1,4 @@
-package keyscmd
+package keyscli
 
 import (
 	"flag"
