+++ gno.land/pkg/keyscli/verify_test.go
@@ -1,4 +1,4 @@
-package keyscmd
+package keyscli
 
 import (
 	"encoding/hex"
